[Refactor] 장소 관련 모든 아이템 누를때 보내는 값 변경
Place Id : Long -> Int로 타입 전부 변경

+++ app/src/main/java/place/pic/ui/main/place/items/PlaceItemsViewModel.kt
@@ -34,7 +34,7 @@ class PlaceItemsViewModel(
     val totalItemCount: LiveData<Int>
         get() = _totalItemCount
 
-    fun removePlace(placeId: Long) {
+    fun removePlace(placeId: Int) {
         val currentItems = getCurrentPlaceItems().toMutableList()
         val filteredItems = currentItems.filter { it.id != placeId }
         _placeItems.value = filteredItems
